refactor: one abstract class per file
