+++ src/fake/Roc.ts
@@ -1,13 +1,11 @@
 /* tslint:disable max-classes-per-file */
 import { randomBytes } from 'crypto';
 
+import BaseRoc from '../base/BaseRoc';
+import BaseRocDocument from '../base/BaseRocDocument';
+import BaseRocQuery from '../base/BaseRocQuery';
+import { BaseRocReduceQuery } from '../base/BaseRocReduceQuery';
 import { RocClientError, RocHTTPError } from '../Error';
-import {
-  BaseRoc,
-  BaseRocDocument,
-  BaseRocQuery,
-  BaseRocReduceQuery
-} from '../RocBase';
 import {
   Encoding,
   IAttachment,
